refactor: authOptions와 refreshAccessToken 함수를 lib 폴더로 분리
